% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.get.excel.R
\name{xl.get.excel}
\alias{xl.get.excel}
\title{Returns reference to Excel application.}
\usage{
xl.get.excel()
}
\value{
object of class 'COMIDispatch' (as returned by 
\code{COMCreate} from RDCOMClient package).
}
\description{
Returns reference to Microsoft Excel application. If there is no running 
instance exists it will create a new instance.
}
\examples{

\dontrun{
xls = xl.get.excel() 
}
}
